requirements.txt
apturl==0.5.2
argcomplete==1.8.1
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
cachetools==5.5.2
certifi==2020.6.20
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
constantly==23.10.4
cryptography==44.0.2
cupshelpers==1.0
daphne==4.1.2
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.9
distro==1.7.0
distro-info==1.1+ubuntu0.2
dj-database-url==2.3.0
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
duplicity==0.8.21
fasteners==0.14.1
filelock==3.16.1
future==0.18.2
google-api-core==2.24.1
google-api-python-client==2.162.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.68.0
gunicorn==23.0.0
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==24.7.2
jeepney==0.7.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
msgpack==1.1.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
packaging==24.2
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
platformdirs==4.3.6
proto-plus==1.26.0
protobuf==5.29.3
psycopg2==2.9.10
ptyprocess==0.7.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycairo==1.20.1
pycparser==2.22
pycups==2.0.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==25.0.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
redis==5.2.1
reportlab==3.6.8
requests==2.25.1
requests-oauthlib==2.0.0
rsa==4.9
sdkmanager==0.5.1
SecretStorage==3.3.1
service-identity==24.2.0
six==1.16.0
sqlparse==0.5.3
ssh-import-id==5.11
systemd-python==234
tomli==2.2.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.27.0
wadllib==1.3.6
xdg==5
xkit==0.0.0
zipp==1.0.0
zope.interface==7.2
